GPSCAFFOLD-85 - improve the default controller spec
The default generated controller uses request.withFormat and the default generated spec isn't compatible with that.

Note that this change requires Grails 2.3.6 or later for a couple reasons.  One is that it is using the FORM_CONTENT_TYPE constant and also because it relies on a code change to allow the contentType to be properly respected when unit tests are run.  See http://jira.grails.org/browse/GRAILS-11093.